#151 replaced long casts in integration test with literals

+++ jdbc-adapter/virtualschema-jdbc-adapter/src/test/java/com/exasol/adapter/dialects/postgresql/PostgreSQLSqlDialectIT.java
@@ -236,7 +236,7 @@ class PostgreSQLSqlDialectIT extends AbstractIntegrationTest {
         final String query = "SELECT * FROM  " + VIRTUAL_SCHEMA + ".t1 a INNER JOIN  " + VIRTUAL_SCHEMA
                 + ".t2 b ON a.x=b.x";
         final ResultSet result = executeQuery(query);
-        assertAll(() -> matchNextRow(result, (long) 2, "bbb", (long) 2, "bbb"), () -> assertFalse(result.next()));
+        assertAll(() -> matchNextRow(result, 2L, "bbb", 2L, "bbb"), () -> assertFalse(result.next()));
     }
 
     @Test
@@ -252,8 +252,8 @@ class PostgreSQLSqlDialectIT extends AbstractIntegrationTest {
         final String query = "SELECT * FROM  " + VIRTUAL_SCHEMA + ".t1 a LEFT OUTER JOIN  " + VIRTUAL_SCHEMA
                 + ".t2 b ON a.x=b.x ORDER BY a.x";
         final ResultSet result = executeQuery(query);
-        matchNextRow(result, (long) 1, "aaa", null, null);
-        matchNextRow(result, (long) 2, "bbb", (long) 2, "bbb");
+        matchNextRow(result, 1L, "aaa", null, null);
+        matchNextRow(result, 2L, "bbb", 2L, "bbb");
         assertFalse(result.next());
     }
 
@@ -262,8 +262,8 @@ class PostgreSQLSqlDialectIT extends AbstractIntegrationTest {
         final String query = "SELECT * FROM  " + VIRTUAL_SCHEMA + ".t1 a RIGHT OUTER JOIN  " + VIRTUAL_SCHEMA
                 + ".t2 b ON a.x=b.x ORDER BY a.x";
         final ResultSet result = executeQuery(query);
-        assertAll(() -> matchNextRow(result, (long) 2, "bbb", (long) 2, "bbb"),
-                () -> matchNextRow(result, null, null, (long) 3, "ccc"), () -> assertFalse(result.next()));
+        assertAll(() -> matchNextRow(result, 2L, "bbb", 2L, "bbb"),
+                () -> matchNextRow(result, null, null, 3L, "ccc"), () -> assertFalse(result.next()));
     }
 
     @Test
@@ -271,9 +271,9 @@ class PostgreSQLSqlDialectIT extends AbstractIntegrationTest {
         final String query = "SELECT * FROM  " + VIRTUAL_SCHEMA + ".t1 a FULL OUTER JOIN  " + VIRTUAL_SCHEMA
                 + ".t2 b ON a.x=b.x ORDER BY a.x";
         final ResultSet result = executeQuery(query);
-        assertAll(() -> matchNextRow(result, (long) 1, "aaa", null, null),
-                () -> matchNextRow(result, (long) 2, "bbb", (long) 2, "bbb"),
-                () -> matchNextRow(result, null, null, (long) 3, "ccc"), () -> assertFalse(result.next()));
+        assertAll(() -> matchNextRow(result, 1L, "aaa", null, null),
+                () -> matchNextRow(result, 2L, "bbb", 2L, "bbb"),
+                () -> matchNextRow(result, null, null, 3L, "ccc"), () -> assertFalse(result.next()));
     }
 
     @Test
@@ -281,8 +281,8 @@ class PostgreSQLSqlDialectIT extends AbstractIntegrationTest {
         final String query = "SELECT * FROM  " + VIRTUAL_SCHEMA + ".t1 a RIGHT OUTER JOIN  " + VIRTUAL_SCHEMA
                 + ".t2 b ON a.x||a.y=b.x||b.y ORDER BY a.x";
         final ResultSet result = executeQuery(query);
-        assertAll(() -> matchNextRow(result, (long) 2, "bbb", (long) 2, "bbb"),
-                () -> matchNextRow(result, null, null, (long) 3, "ccc"), () -> assertFalse(result.next()));
+        assertAll(() -> matchNextRow(result, 2L, "bbb", 2L, "bbb"),
+                () -> matchNextRow(result, null, null, 3L, "ccc"), () -> assertFalse(result.next()));
     }
 
     @Test
@@ -290,9 +290,9 @@ class PostgreSQLSqlDialectIT extends AbstractIntegrationTest {
         final String query = "SELECT * FROM  " + VIRTUAL_SCHEMA + ".t1 a FULL OUTER JOIN  " + VIRTUAL_SCHEMA
                 + ".t2 b ON a.x-b.x=0 ORDER BY a.x";
         final ResultSet result = executeQuery(query);
-        assertAll(() -> matchNextRow(result, (long) 1, "aaa", null, null),
-                () -> matchNextRow(result, (long) 2, "bbb", (long) 2, "bbb"),
-                () -> matchNextRow(result, null, null, (long) 3, "ccc"), () -> assertFalse(result.next()));
+        assertAll(() -> matchNextRow(result, 1L, "aaa", null, null),
+                () -> matchNextRow(result, 2L, "bbb", 2L, "bbb"),
+                () -> matchNextRow(result, null, null, 3L, "ccc"), () -> assertFalse(result.next()));
     }
 
     // Identifier Test - CONVERT_TO_UPPER mode --------------------------------
